backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
langchain>=0.1.10
langchain-community>=0.0.25
langchain-core>=0.1.28
langchain-openai>=0.0.8
openai>=1.0.0
pymupdf>=1.23.26
python-docx>=1.1.0
faiss-cpu>=1.7.4
pydantic>=2.6.3
pydantic-settings>=2.2.1
httpx>=0.27.0
jinja2>=3.1.3
python-dotenv>=1.0.1
pytest>=8.0.2
pytest-asyncio>=0.23.5
